[WB-2354] ActivityIconButton: fit long translated labels via Body/Small + hyphenation
Implements the label-fitting approach validated by the WB-2354 spike. The
visible label now renders at Body/Small and uses locale-aware hyphenation
(hyphens: auto with wordBreak: normal + overflowWrap: break-word), so longer
international launch strings stay within two lines instead of being clipped.

Replaces the temporary spike stories with a permanent regression snapshot:
the Tier 1 launch locale matrix now lives in the ActivityIconButton testing
snapshots (LabelFittingLocaleMatrix), rendering the real component across 12
locales (incl. RTL + Indic) × 9 controls. Severity dots are accessible
(role="img" + aria-label, contrasting ring for non-text contrast).

Issue: WB-2354
