--- .changeset/wb-2354-label-fitting-spike.md
@@ -1,4 +1,0 @@
----
----
-
-Spike (WB-2354): explore `ActivityIconButton` label-fitting techniques (hyphenation, Body/Small, 2-line wrap) against the real Tier 1 international launch string set. Storybook-only — no published package changes.
