+++ packages/wonder-blocks-icon-button/src/components/activity-icon-button.tsx
@@ -138,7 +138,7 @@ export const ActivityIconButton: React.ForwardRefExoticComponent<
                 {hasVisibleLabel && (
                     <BodyText
                         tag="span"
-                        size="medium"
+                        size="small"
                         weight="semi"
                         style={[buttonStyles.label, stylesProp?.label]}
                     >
@@ -335,7 +335,14 @@ const _generateStyles = (
             // We restrict the label to a maximum of 2 lines.
             WebkitLineClamp: "2",
             overflow: "hidden",
-            wordBreak: "break-word",
+            // Label fitting for long (esp. translated) labels: prefer
+            // locale-aware hyphenation, then fall back to breaking long words
+            // so nothing overflows. `wordBreak: normal` is required so hyphens
+            // can win — `break-word` would force a hard break first.
+            // @see WB-2354
+            hyphens: "auto",
+            wordBreak: "normal",
+            overflowWrap: "break-word",
         },
     } as const;
